Preparations for version 5.4
