+++ appveyor.yml
@@ -1,4 +1,4 @@
-version: 5.3.2.{build}
+version: 5.4.0.{build}
 branches:
   only:
   - develop
@@ -6,15 +6,15 @@ configuration: release
 platform: Any CPU
 init:
 - ps: >-
-    Update-AppveyorBuild -Version "5.3.2.$env:appveyor_build_number-$(Get-Date -format yyyyMMdd)-$env:appveyor_repo_branch"
+    Update-AppveyorBuild -Version "5.4.0.$env:appveyor_build_number-$(Get-Date -format yyyyMMdd)-$env:appveyor_repo_branch"
 
-    Write-Host "5.3.2.$env:appveyor_build_number-$(Get-Date -format yyyyMMdd)-$env:appveyor_repo_branch"
+    Write-Host "5.4.0.$env:appveyor_build_number-$(Get-Date -format yyyyMMdd)-$env:appveyor_repo_branch"
 dotnet_csproj:
   patch: true
   file: '**\*.csproj'
   version: '{version}'
-  assembly_version: 5.3.2.{build}
-  file_version: 5.3.2.{build}
+  assembly_version: 5.4.0.{build}
+  file_version: 5.4.0.{build}
 nuget:
   project_feed: true
 before_build:
