+++ src/EPPlus/Drawing/Slicer/ExcelSlicer.cs
@@ -17,24 +17,10 @@ using System.Xml;
 
 namespace OfficeOpenXml.Drawing.Slicer
 {
-    /*
-      <xsd:complexType name="CT_Slicer">
-       <xsd:sequence>
-         <xsd:element name="extLst" type="x:CT_ExtensionList" minOccurs="0" maxOccurs="1"/>
-       </xsd:sequence>
-       <xsd:attribute name="name" type="x:ST_Xstring" use="required"/>
-       <xsd:attribute ref="xr10:uid" use="optional"/>
-       <xsd:attribute name="cache" type="x:ST_Xstring" use="required"/>
-       <xsd:attribute name="caption" type="x:ST_Xstring" use="optional"/>
-       <xsd:attribute name="startItem" type="xsd:unsignedInt" use="optional" default="0"/>
-       <xsd:attribute name="columnCount" type="xsd:unsignedInt" use="optional" default="1"/>
-       <xsd:attribute name="showCaption" type="xsd:boolean" use="optional" default="true"/>
-       <xsd:attribute name="level" type="xsd:unsignedInt" use="optional" default="0"/>
-       <xsd:attribute name="style" type="x:ST_Xstring" use="optional"/>
-       <xsd:attribute name="lockedPosition" type="xsd:boolean" use="optional" default="false"/>
-       <xsd:attribute name="rowHeight" type="xsd:unsignedInt" use="required"/>
-     </xsd:complexType>
-     */
+    /// <summary>
+    /// Base class for table and pivot table slicers.
+    /// </summary>
+    /// <typeparam name="T">The slicer cache data type</typeparam>
     public abstract class ExcelSlicer<T> : ExcelDrawing where T : ExcelSlicerCache
     {
         internal ExcelSlicerXmlSource _xmlSource;
@@ -233,6 +219,9 @@ namespace OfficeOpenXml.Drawing.Slicer
             }
         }
         protected internal ExcelSlicerCache _cache = null;
+        /// <summary>
+        /// A reference to the slicer cache.
+        /// </summary>
         public T Cache
         {
             get
