+++ src/EPPlus/Table/PivotTable/ExcelPivotTableDataField.cs
@@ -21,7 +21,7 @@ using OfficeOpenXml.Style.XmlAccess;
 namespace OfficeOpenXml.Table.PivotTable
 {
     /// <summary>
-    /// A pivo table data field
+    /// A pivot table data field
     /// </summary>
     public class ExcelPivotTableDataField : XmlHelper
     {
